Logs

+++ cmds/Info/say.js
@@ -10,15 +10,15 @@ module.exports = {
     run: async (client, message, args) => {
     let argsresult;
     let mChannel = message.mentions.channels.first();
-    if (message.guild.me.hasPermission("MANAGE_MESSAGES" || "ADMINISTRATOR")) {
-    message.delete()
-    }
+    if (message.guild.me.hasPermission("MANAGE_MESSAGES" || "ADMINISTRATOR")) message.delete()
+
     if(mChannel && args[0] == "{sendchannel}") {
+        if (mChannel.permissionsFor(message.guild.me).has("SEND_MESSAGES")) return
         argsresult = args.join(" ")
         if(!argsresult) return;
         if(!message.member.hasPermission("MANAGE_MESSAGES" || "ADMINISTRATOR") || !message.guild.owner) {
             mChannel = message.channel
-           let replace = argsresult.replace(mChannel, '').replace(args[0], '').replace("@everyone", "everyone").replace("@here", "here").replace(/(https?:\/\/)?(www\.)?(discord\.(gg|io|me|li|com)|discordapp\.com\/invite)\/.+[A-z0-9]/, "")
+           let replace = argsresult.replace(mChannel, '').replace(args[0], '').replace(/@everyone/g, "everyone").replace(/@here/g, "here").replace(/(https?:\/\/)?(www\.)?(discord\.(gg|io|me|li|com)|discordapp\.com\/invite)\/.+[A-z0-9]/, "")
             mChannel.send(replace)
         }
        else {
@@ -29,7 +29,7 @@ module.exports = {
         argsresult = args.join(" ")
         if(!argsresult) return;
         if(!message.member.hasPermission("MANAGE_MESSAGES" || "ADMINISTRATOR") || !message.guild.owner) {
-               let replace = argsresult.replace("@everyone", "everyone").replace("@here", "here").replace(/(https?:\/\/)?(www\.)?(discord\.(gg|io|me|li|com)|discordapp\.com\/invite)\/.+[A-z0-9]/, "")
+               let replace = argsresult.replace(/@everyone/g, "everyone").replace(/@here/g, "here").replace(/(https?:\/\/)?(www\.)?(discord\.(gg|io|me|li|com)|discordapp\.com\/invite)\/.+[A-z0-9]/, "")
                 message.channel.send(replace)
             } else  message.channel.send(argsresult)
         }
